Add in images downloaded from the Contentful CDN, for local serving instead
Some of the filenames are not very elegant, but they match what we are
using in the templates, which draw data fro the DB, so if we rename them
as files, here we should rename them in the DB too. They will only be around
for a relatively short time as we move them back to a CMS before long.